changing currency long to integer

+++ mscurrency/src/main/java/br/com/mscurrency/dto/CreateCurrencyPriceInput.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class CreateCurrencyPriceInput {
     private String currencyCode;
-    private Long priceInCentsBRL;
+    private Integer priceInCentsBRL;
 }
