+++ mscurrency/src/main/java/br/com/mscurrency/services/CurrencyPriceService.java
@@ -8,8 +8,8 @@ import java.util.Optional;
 public interface CurrencyPriceService {
     List<CurrencyPrice> findAll();
     Optional<CurrencyPrice> findByCurrencyCode(String currencyCode);
-    CurrencyPrice createCurrencyPrice(String currencyCode, Long priceInCentsBRL);
-    CurrencyPrice updateCurrencyPrice(String currencyCode, Long priceInCentsBRL);
+    CurrencyPrice createCurrencyPrice(String currencyCode, Integer priceInCentsBRL);
+    CurrencyPrice updateCurrencyPrice(String currencyCode, Integer priceInCentsBRL);
     boolean deleteCurrencyPrice(String currencyCode);
     boolean existsByCurrencyCode(String currencyCode);
 }
